+++ src/app/components/contact/contact.component.html
@@ -1,0 +1,12 @@
+<h2 class="header">{{ header }}</h2>
+<div class="contact-list" *ngFor="let contact of contactMethods">
+  <div class="box" (click)="openLink(contact)">
+    <div class="contact-info">
+      <h2>{{ contact.media }}</h2>
+      <div class="contact-details">
+        <p>{{ contact.username }}</p>
+      </div>
+    </div>
+    <img class="img" src="{{ contact.image_path }}" />
+  </div>
+</div>
