+++ src/app/components/content-section/content-section.component.css
@@ -8,7 +8,7 @@
   display: grid;
   grid-template-columns: 1.7fr 0.3fr;
   justify-content: space-between;
-  transition: max-width 1s, max-height 1s;
+  transition: max-width 0.7s, max-height 0.7s;
   overflow-wrap: break-word;
   overflow: hidden;
 }
@@ -51,14 +51,14 @@
 .img {
   max-width: 20vh;
   height: auto;
-  transition: max-width 1s;
+  transition: max-width 0.7s;
 }
 
 .details {
   opacity: 0%;
   max-height: 0%;
   max-height: 0%;
-  transition: opacity 1s ease-in-out;
+  transition: opacity 0.7s ease-in-out;
 }
 
 .box:hover .details {
