--- src/app/contact/contact.component.html
@@ -1,11 +1,0 @@
-<div class="contact-list" *ngFor="let contact of contactMethods">
-    <div class="box" (click)="openLink(contact)">
-        <div class="contact-info">
-            <h2>{{contact.media}}</h2>
-            <div class="contact-details">
-                <p>{{contact.username}}</p>
-            </div>
-        </div>
-        <img class="img" src={{contact.image_path}}>
-    </div>
-</div>
