--- src/app/education/education.component.html
@@ -1,20 +1,0 @@
-<div class="box">
-    <div class="education-info">
-        <h2>Kansas State University</h2>
-        <h4>B.S. Computer Science</h4>
-        <h4>Aug 2019 - May 2023</h4>
-        <p>Manhattan, KS</p>
-        <div class="education-details">
-            <p>Relevant Coursework</p> 
-            <p>- Data and Program Structures</p>
-            <p>- Logical Foundations of Programming</p>
-            <p>- Object-Oriented Design</p>
-            <p>- Implementation and Testing</p>
-            <p>- Ethics and Conduct for Computing Professionals</p>
-            <p>Minor - Japanese</p>
-        </div>
-    </div>
-    <div class="img_box">
-        <img class="img" src=assets\img\ksulogo.png>
-    </div>
-</div>
