--- src/app/experience/experience.component.css
@@ -1,60 +1,0 @@
-.box {
-    color: black;
-    background-color: #F4F7F5;
-    padding: 1% 1%;
-    margin: 1% 1%;
-    max-height: 15vh;
-    max-width: 70%;
-    display: flex;
-    justify-content: space-between;
-    transition: max-width 1s, max-height 1s;
-    overflow: hidden;
-
-}
-
-.box:hover {
-    max-width: 95%;
-    max-height: 50vh;
-} 
-
-.box p {
-    font-size: 2vh;
-    padding: 1%;
-}
-
-.box h2 {
-    font-size: 3vh;
-    padding: 1%;
-}
-
-.box h4 {
-    font-size: 2vh;
-    padding: 1%;
-}
-
-.experience-info {
-    display: grid;
-    padding-right: 10%;
-    margin-left: 1%;
-}
-
-.img {
-    max-width: 25vh;
-    height: auto;
-    border-radius: 10px;
-    display: block;
-}
-
-.experience-details {
-    opacity: 0%;
-    max-height: 0%;
-    background-color: aqua;
-    max-height: 0%;
-    transition: opacity 1s ease-in-out;
-    margin-left: 2%;
-}
-
-.box:hover .experience-details {
-    opacity: 100%;
-}
-
