--- src/app/experience/experience.component.html
@@ -1,19 +1,0 @@
-<div class="box">
-    <div class="experience-info">
-        <h2>Developer Intern - WellSky</h2>
-        <h4>Jun 2022 - Aug 2022</h4>
-        <p>Overland Park, KS</p>
-        <div class="experience-details">
-            <p>Contributed to world-class blood transfusion application</p> 
-            <p>- Accomplished goals using Agile and Scrum frameworks </p>
-            <p>- Created and expanded UI components using Angular Framework </p>
-            <p>- Expanded and integrated APIs using .NET Core Framework </p>
-            <p>- Deployed pipelines using Azure </p>
-            <p>- Assisted in porting legacy version to modern frameworks </p>
-            <p>- Applied multilingual skills to collaborate with South America-based colleagues</p>
-        </div>
-    </div>
-    <div class="img_box">
-        <img class="img" src=assets\img\wellskylogo.png>
-    </div>
-</div>
