--- src/app/experience/experience.component.spec.ts
@@ -1,21 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { ExperienceComponent } from './experience.component';
-
-describe('ExperienceComponent', () => {
-  let component: ExperienceComponent;
-  let fixture: ComponentFixture<ExperienceComponent>;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      declarations: [ExperienceComponent]
-    });
-    fixture = TestBed.createComponent(ExperienceComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
