--- src/app/mock-api.ts
@@ -1,18 +1,0 @@
-import { Project } from './project';
-
-export const CONTENT: Project[] = [
-  {
-    id: 1,
-    name: 'Quictionary',
-    summary: 'A powerful flashcard study application',
-    techstack: 'Swift, Core Data',
-    cover_image_path: '/assets/img/quictionary-cover-image.png',
-  },
-  {
-    id: 2,
-    name: 'Project Name',
-    summary: 'Project Summary',
-    techstack: 'Project Tech Stack',
-    cover_image_path: '/assets/img/quictionary-cover-image.png',
-  },
-];
