--- src/app/nav/nav.component.css
@@ -1,61 +1,0 @@
-.navbar {
-    height: 100vh;
-    width: 100%;
-    color: #F4F7F5;
-    overflow: visible;
-}
-
-.site-menu {
-    font-size: 1vw;
-    width: 100%;
-    padding-left: 2%;
-    text-align: center;
-    line-height: 15vh;
-    display: grid;
-}
-
-.site-menu a {
-    text-decoration: none;
-    margin: 1% 0%;
-    /* color: #F4F7F5; */
-    color: #04030F;
-    height: 15vh;
-/*     border-radius: 80px; */
-    /* clip-path: polygon(25% 0%, 100% 0%, 75% 100%, 0% 100%); */
-    clip-path: polygon(0% 0%, 100% 0%, 100% 100%, 0% 100%);
-    /* background-color: #2B3A67; */
-    background-color: #E8E9EB;
-    width: 50%;
-    font-size: 2vw;
-    text-shadow: 0px 0px black;
-    transition: font-size 0.5s, width 1s, padding 1s, clip-path 1s, color 1s, text-shadow 1s;
-    position: relative;
-}
-
-.site-menu a:hover {
-    font-size: 2.5vw;
-    width: 60%;
-    clip-path: polygon(25% 0%, 100% 0%, 75% 100%, 0% 100%);
-}
-
-.navButton-wrap:hover {
-    transition: filter 0.2s;
-    filter: drop-shadow(30px 10px 0px #06BA63);
-}
-
-.navButton-wrap {
-    display: flex;
-}
-
-.navButton-wrap.active {
-    filter: drop-shadow(50px 20px 0px #06BA63);
-    z-index: -1;
-}
-
-.site-menu a.active {
-    width: 80%;
-    font-size: 3vw;
-    clip-path: polygon(25% 0%, 100% 0%, 75% 100%, 0% 100%);
-    pointer-events: auto;
-}
-
