--- src/app/project.service.ts
@@ -1,26 +1,0 @@
-import { Injectable } from '@angular/core';
-import { HttpClient } from '@angular/common/http';
-import { Project } from './project';
-import { Observable, catchError, of, tap } from 'rxjs';
-
-@Injectable({
-  providedIn: 'root',
-})
-export class ProjectService {
-  private projectsUrl = 'api/projects';
-
-  constructor(private http: HttpClient) {}
-
-  getProjects(): Observable<Project[]> {
-    return this.http
-      .get<Project[]>(this.projectsUrl)
-      .pipe(tap(), catchError(this.handleError<Project[]>('getProjects', [])));
-  }
-
-  private handleError<T>(operation = 'operation', result?: T) {
-    return (error: any): Observable<T> => {
-      console.error(error);
-      return of(result as T);
-    };
-  }
-}
