--- src/app/project.ts
@@ -1,8 +1,0 @@
-export interface Project {
-    id: number,
-    name: string;
-    summary?: string;
-    techstack?: string;
-    link?: string;
-    cover_image_path: string;
-}
