--- src/app/project/project.component.css
@@ -1,62 +1,0 @@
-.box {
-    color: black;
-    background-color: #F4F7F5;
-    padding: 1% 1%;
-    margin: 1% 1%;
-    max-height: 15vh;
-    max-width: 70%;
-    display: flex;
-    justify-content: space-between;
-    transition: max-width 1s, max-height 1s;
-    overflow: hidden;
-}
-
-.box:hover {
-    max-width: 95%;
-    max-height: 50vh;
-}
-
-.box p {
-    font-size: 2vh;
-    padding: 1%;
-}
-
-.box h2 {
-    font-size: 3vh;
-    padding: 1%;
-}
-
-.box h4 {
-    font-size: 2vh;
-    padding: 1%;
-}
-
-.project-info {
-    display: grid;
-    margin-left: 1%;
-}
-
-.img {
-    max-width: 15vh;
-    /* max-height: auto; */
-    transition: max-width 1s, max-height 1s;
-}
-
-.project-details {
-    opacity: 0%;
-    max-height: 0%;
-    max-height: 0%;
-    transition: opacity 1s ease-in-out;
-}
-
-.box:hover .project-details {
-    opacity: 100%;
-}
-
-.box:hover .img {
-    max-width: 30vh;
-    max-height: auto;
-}
-
-
-
