--- src/app/project/project.component.html
@@ -1,13 +1,0 @@
-<div class="box">
-    <div class="project-info" (click)="openLink(proj)">
-        <h2>{{proj.name}}</h2>
-        <p>{{proj.techstack}}</p>
-        <div class="project-details">
-            <p>{{proj.summary}}</p>
-        </div>
-    </div>
-    <div class="img_box">
-        <img class="img" src={{proj.cover_image_path}}>
-    </div>
-</div>
-
