+++ src/app/services/contact.service.ts
@@ -1,7 +1,6 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
-import { Project } from './project';
-import { Contact } from './contact';
+import { Contact } from '../models/contact';
 import { Observable, catchError, of, tap } from 'rxjs';
 
 @Injectable({
