+++ src/app/services/content.service.ts
@@ -1,7 +1,6 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
-import { Project } from './project';
-import { Content } from './content';
+import { Content } from '../models/content';
 import { Observable, catchError, of, tap } from 'rxjs';
 
 @Injectable({
