+++ src/app/services/in-memory-data.service.ts
@@ -1,6 +1,5 @@
 import { Injectable } from '@angular/core';
 import { InMemoryDbService } from 'angular-in-memory-web-api';
-import { Project } from './project';
 
 @Injectable({
   providedIn: 'root',
@@ -85,11 +84,5 @@ export class InMemoryDataService implements InMemoryDbService {
     ];
     const navItems = ['Projects', 'Experience', 'Education', 'Contact'];
     return { content, contactMethods, navItems };
-  }
-
-  genId(projects: Project[]): number {
-    return projects.length > 0
-      ? Math.max(...projects.map((project) => project.id)) + 1
-      : 11;
   }
 }
